+++ lib/runners/workspace/git.rb
@@ -2,11 +2,14 @@ module Runners
   class Workspace::Git < Workspace
     class FetchFailed < SystemError; end
     class CheckoutFailed < SystemError; end
+    class BlameFailed < SystemError; end
 
     def range_git_blame_info(path_string, start_line, end_line)
       stdout, _ = shell.capture3!("git", "blame", "-p", "-L", "#{start_line},#{end_line}", git_source.head, "--", path_string,
                                   trace_stdout: false, trace_stderr: true)
       GitBlameInfo.parse(stdout)
+    rescue Shell::ExecError => exn
+      raise BlameFailed, "git-blame failed: #{exn.stderr_str}"
     end
 
     private